extract component and add props for custom show
